+++ finance/src/main/kotlin/net/corda/finance/flows/TwoPartyDealFlow.kt
@@ -7,8 +7,8 @@ import net.corda.core.crypto.TransactionSignature
 import net.corda.core.flows.*
 import net.corda.core.identity.AnonymousParty
 import net.corda.core.identity.Party
-import net.corda.core.internal.excludeNotary
-import net.corda.core.internal.groupPublicKeysByWellKnownParty
+import net.corda.core.identity.excludeNotary
+import net.corda.core.identity.groupPublicKeysByWellKnownParty
 import net.corda.core.serialization.CordaSerializable
 import net.corda.core.transactions.SignedTransaction
 import net.corda.core.transactions.TransactionBuilder
